+++ src/CleanRegex/Builder/TemplateBuilder.php
@@ -1,7 +1,6 @@
 <?php
 namespace TRegx\CleanRegex\Builder;
 
-use TRegx\CleanRegex\Internal\Prepared\Condition\PositiveCondition;
 use TRegx\CleanRegex\Internal\Prepared\Expression\Template;
 use TRegx\CleanRegex\Internal\Prepared\Orthography\Orthography;
 use TRegx\CleanRegex\Internal\Prepared\Template\AlternationToken;
@@ -46,6 +45,6 @@ class TemplateBuilder
 
     public function build(): Pattern
     {
-        return new Pattern(new Template($this->orthography->spelling(new PositiveCondition()), $this->tokens->figures()));
+        return new Pattern(new Template($this->orthography->spelling($this->tokens->condition()), $this->tokens->figures()));
     }
 }