+++ src/CleanRegex/Internal/Prepared/Template/DelimiterAgnostic.php
@@ -1,0 +1,10 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Prepared\Template;
+
+trait DelimiterAgnostic
+{
+    public function suitable(string $candidate): bool
+    {
+        return true;
+    }
+}