+++ src/CleanRegex/Internal/Prepared/Template/Token.php
@@ -1,10 +1,11 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Prepared\Template;
 
+use TRegx\CleanRegex\Internal\Prepared\Condition\Condition;
 use TRegx\CleanRegex\Internal\Prepared\Quotable\Quotable;
 use TRegx\CleanRegex\Internal\Type;
 
-interface Token
+interface Token extends Condition
 {
     public function formatAsQuotable(): Quotable;