+++ src/CleanRegex/Internal/Prepared/Tokens.php
@@ -1,6 +1,8 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Prepared;
 
+use TRegx\CleanRegex\Internal\Prepared\Condition\CompositeCondition;
+use TRegx\CleanRegex\Internal\Prepared\Condition\Condition;
 use TRegx\CleanRegex\Internal\Prepared\Figure\CountedFigures;
 use TRegx\CleanRegex\Internal\Prepared\Figure\TokenFigures;
 use TRegx\CleanRegex\Internal\Prepared\Template\Token;
@@ -18,6 +20,11 @@ class Tokens
     public function next(Token $token): Tokens
     {
         return new Tokens(\array_merge($this->tokens, [$token]));
+    }
+
+    public function condition(): Condition
+    {
+        return new CompositeCondition($this->tokens);
     }
 
     public function figures(): CountedFigures