+++ test/Utils/Impl/ThrowToken.php
@@ -7,6 +7,11 @@ use TRegx\CleanRegex\Internal\Type;
 
 class ThrowToken implements Token
 {
+    public function suitable(string $candidate): bool
+    {
+        throw new \AssertionError("Token wasn't supposed to be used");
+    }
+
     public function formatAsQuotable(): Quotable
     {
         throw new \AssertionError("Token wasn't supposed to be used");