+++ ChangeLog.md
@@ -5,8 +5,9 @@ Incoming
 --------
 
 * Bug fixes
-    * Fixed w bug where using `match()->filter()` didn't throw `InvalidReturnValueException`.
-    * Fixed w bug where using `group()->filter()` didn't throw `InvalidReturnValueException`.
+    * Fixed a bug, where using `match()->filter()` didn't throw `InvalidReturnValueException`.
+    * Fixed a bug, where using `group()->filter()` didn't throw `InvalidReturnValueException`.
+    * Fixed a bug, where `Pattern::template()->mask()` keywords weren't taken into account, when choosing a delimiter.
 
 Added in 0.13.5
 ---------------